requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.36.0
peft>=0.7.0
timm>=0.9.0
wandb>=0.16.0
tqdm
numpy
Pillow
